requirements.txt
pandas
scikit-learn
pyyaml
dvc
mlflow
joblib
numpy